\name{holford}
\docType{data}
\alias{holford}
\title{Holford's data on prostatic cancer deaths}
\usage{data(minitab)}
\description{
Holford's data on prostatic cancer deaths and mid-period population denominators for non-whites in the US by age and calendar period. Thirteen birth cohorts from 1855-59 through to 1915-19 are represented in at least one of seven 5-year age groups (50-54 through to 80-84) and one of the seven 5-year calendar periods (1935-39 through to 1965-69) for which data are provided. 

}
\format{
This data frame contains the following columns:
\describe{
  \item{incid}{number ofd prostatic cancer deaths.} 
  \item{pop}{mid-period population counts.}
  \item{Age}{age groups.}
  \item{Period}{calendar periods.}
  \item{Cohort}{cohorts.}
}
}
%\details{}
\source{Holford, T.R. (1983) The estimation of age, period and cohort effects for vital rates. \emph{Biometrics}, \bold{39}, 311--324.}
\references{ Breslow, N.E. (1984), Extra-Poisson variation in log-linear models, \emph{Applied Statistics}, \bold{33}, 38--44.}
%\examples{}
\keyword{datasets}
