\name{minitab}
\docType{data}
\alias{minitab}
\title{Minitab tree data}
\usage{data(minitab)}
\description{
Data on 31 black cherry trees sampled from the Allegheny Natinoal Forest,
Pennsylvania.  
}
\format{
This data frame contains the following columns:
\describe{
  \item{D}{diameter 4.5 feet of the ground, inches}
  \item{H}{height of the tree, feet}
  \item{V}{marketable volume of wood, cubic feet}
}
}
\details{}
\source{Ryan, T.A., Joiner, B.L. and Ryan, B.F. (1976) \emph{Minitab Student Handbook}. N. Scituate, MA: Duxbury.}
\references{Cook, R.D. and Weisberg, S. (1982) \emph{Residuals and Influence in Regression}, New York:  Chapman and Hall, p. 66.}
\examples{}
\keyword{datasets}
