\name{orobanche}
\alias{orobanche}
\non_function{}
\title{Germination of Orobanche}
\usage{data(orobanche)}
\description{
Orobanche, commonly known as broomrape, is a genus of parasitic plants with
chlorophyll that grow on the roots of flowering plants.  Batches of seeds of
two varieties of the plant were were brushed onto a plate of diluted extract
of bean or cucumber, and the number germinating were recorded.
}
\format{
This data frame contains the following columns:
\describe{
  \item{germinated}{Number germinated}
  \item{seeds}{Number of seeds}
  \item{slide}{Slide number}
  \item{host}{Host type}
  \item{variety}{Variety name}
}
}
\details{}
\source{Crowder, M.J. (1978) Beta-binomial anova for proportions. \emph{Applied Statistics}, \bold{27}, 34--37.}

\references{Collett, D. (1991) \emph{Modelling Binary Data}, London: Chapman and Hall, Chapter 6.}
\examples{}
\keyword{datasets}
