\name{salmonellaTA98}
\docType{data}
\alias{salmonellaTA98}
\title{Salmonella reverse mutagenicity assay}
\usage{data(salmonellaTA98)}
\description{
Data on Ames Salmonella reverse mutagenicity assay.
}
\format{
This data frame contains the following columns:
\describe{
  \item{x}{dose levels of quinoline}
  \item{y}{numbers of revertant colonies of TA98 Salmonella observed on each of three replicate plates testes at each of six dose levels of quinolinediameter 4.5 feet of the ground, inches}
}
}
%\details{}
\source{Margolin, B.J., Kaplan, N. and Zeiger, E. (1981) Statistical analysis of the Ames Salmonella/microsome test, \emph{Proc. Natl. Acad. Sci. USA}, \bold{76}, 3779--3783.}
\references{ Breslow, N.E. (1984), Extra-Poisson variation in log-linear models, \emph{Applied Statistics}, \bold{33}, 38--44.}
%\examples{}
\keyword{datasets}
